+++ lib/screens/tools/timestamp_camera/lens_selector_widget.dart
@@ -5,6 +5,7 @@ class LensSelectorWidget extends StatelessWidget {
   final double currentZoom;
   final double minZoom;
   final double maxZoom;
+  final bool hasUltraWide;
   final ValueChanged<double> onZoomChanged;
 
   const LensSelectorWidget({
@@ -12,6 +13,7 @@ class LensSelectorWidget extends StatelessWidget {
     required this.currentZoom,
     required this.minZoom,
     required this.maxZoom,
+    this.hasUltraWide = false,
     required this.onZoomChanged,
   });
 
@@ -19,7 +21,7 @@ class LensSelectorWidget extends StatelessWidget {
   Widget build(BuildContext context) {
     final stops = <_LensStop>[];
 
-    if (minZoom <= 0.6) {
+    if (hasUltraWide || minZoom <= 0.6) {
       stops.add(_LensStop(label: '.5', zoom: 0.5));
     }
     stops.add(_LensStop(label: '1x', zoom: 1.0));